Remove unused space filling + move text inputs previews to common source set
