--- carbon/src/androidDebug/kotlin/com/gabrieldrn/carbon/textinput/TextAreaPreview.kt
@@ -1,93 +1,0 @@
-/*
- * Copyright 2024 Gabriel Derrien
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.gabrieldrn.carbon.textinput
-
-import androidx.compose.foundation.layout.padding
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.tooling.preview.PreviewParameter
-import com.gabrieldrn.carbon.CarbonDesignSystem
-import com.gabrieldrn.carbon.foundation.spacing.SpacingScale
-
-@Preview
-@Composable
-private fun EmptyTextAreaPreview(
-) {
-    var text by remember {
-        mutableStateOf("")
-    }
-
-    CarbonDesignSystem {
-        TextArea(
-            label = "Label",
-            value = text,
-            onValueChange = { text = it },
-            modifier = Modifier.padding(SpacingScale.spacing03),
-            placeholderText = "Placeholder",
-            maxLines = 7
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun TextAreaPreview(
-    @PreviewParameter(TextInputStatePreviewParameterProvider::class) state: TextInputState
-) {
-    var text by remember {
-        mutableStateOf(loremIpsum)
-    }
-
-    CarbonDesignSystem {
-        TextArea(
-            label = "Label",
-            value = text,
-            onValueChange = { text = it },
-            modifier = Modifier.padding(SpacingScale.spacing03),
-            placeholderText = "Placeholder",
-            helperText = state.name,
-            state = state,
-            maxLines = 7
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun TextAreaCounterPreview() {
-    var text by remember {
-        mutableStateOf(loremIpsum)
-    }
-
-    CarbonDesignSystem {
-        TextArea(
-            label = "Label",
-            value = text,
-            onValueChange = { text = it },
-            modifier = Modifier.padding(SpacingScale.spacing03),
-            placeholderText = "Placeholder",
-            helperText = "With counter",
-            counter = text.split(" ").size to 100,
-            maxLines = 7
-        )
-    }
-}
