+++ carbon/src/commonMain/kotlin/com/gabrieldrn/carbon/textinput/PasswordInput.kt
@@ -18,6 +18,7 @@ package com.gabrieldrn.carbon.textinput
 
 import androidx.compose.foundation.interaction.Interaction
 import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
@@ -27,6 +28,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.referentialEqualityPolicy
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.SolidColor
 import androidx.compose.ui.graphics.vector.ImageVector
@@ -38,13 +40,17 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import com.gabrieldrn.carbon.Carbon
+import com.gabrieldrn.carbon.CarbonDesignSystem
 import com.gabrieldrn.carbon.button.ButtonType
 import com.gabrieldrn.carbon.button.IconButton
 import com.gabrieldrn.carbon.common.semantics.imageVectorName
 import com.gabrieldrn.carbon.foundation.color.LocalCarbonTheme
 import com.gabrieldrn.carbon.foundation.color.Theme
+import com.gabrieldrn.carbon.foundation.spacing.SpacingScale
 import com.gabrieldrn.carbon.icons.viewIcon
 import com.gabrieldrn.carbon.icons.viewOffIcon
+import org.jetbrains.compose.ui.tooling.preview.Preview
+import org.jetbrains.compose.ui.tooling.preview.PreviewParameter
 
 /**
  * A modified [KeyboardOptions.Default] with autoCorrect set to false and keyboardType set to
@@ -193,6 +199,33 @@ private fun HidePasswordButton(
             modifier = modifier
                 .semantics { imageVectorName(icon.name) }
                 .testTag(TextInputTestTags.HIDE_PASSWORD_BUTTON)
+        )
+    }
+}
+
+@Preview
+@Composable
+private fun PasswordInputPreview(
+    @PreviewParameter(TextInputStatePreviewParameterProvider::class) state: TextInputState
+) {
+    var text by remember {
+        mutableStateOf("S0mePa55word%")
+    }
+    var passwordHidden by remember {
+        mutableStateOf(false)
+    }
+
+    CarbonDesignSystem {
+        PasswordInput(
+            label = "Label",
+            value = text,
+            passwordHidden = passwordHidden,
+            onValueChange = { text = it },
+            onPasswordHiddenChange = { passwordHidden = it },
+            modifier = Modifier.padding(SpacingScale.spacing03),
+            placeholderText = "Placeholder",
+            helperText = state.name,
+            state = state,
         )
     }
 }
